# [v6-exp] Expand trace support for SSH and Performance