--- a/_examples/common_test.go
+++ b/_examples/common_test.go
@@ -25,6 +25,7 @@ var args = map[string][]string{
 	"find-if-any-tag-point-head": {cloneRepository(defaultURL, tempFolder())},
 	"ls":                         {cloneRepository(defaultURL, tempFolder()), "HEAD", "vendor"},
 	"ls-remote":                  {defaultURL},
+	"memory":                     {defaultURL},
 	"merge_base":                 {cloneRepository(defaultURL, tempFolder()), "--is-ancestor", "HEAD~3", "HEAD^"},
 	"open":                       {cloneRepository(defaultURL, tempFolder())},
 	"progress":                   {defaultURL, tempFolder()},