--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,11 @@ module github.com/go-git/go-git/v5
 // go-git supports the last 3 stable Go versions.
 go 1.20
 
-replace github.com/go-git/go-billy/v5 => github.com/go-git/go-billy/v5 v5.0.0-20240804231525-dc481f5289ba
+// Use the v6-exp branch across go-git dependencies (gcfg and go-billy).
+replace (
+	github.com/go-git/gcfg => github.com/go-git/gcfg v1.5.1-0.20240812080926-1b398f6213c9
+	github.com/go-git/go-billy/v5 => github.com/go-git/go-billy/v5 v5.0.0-20240804231525-dc481f5289ba
+)
 
 require (
 	dario.cat/mergo v1.0.0
@@ -43,6 +47,5 @@ require (
 	golang.org/x/mod v0.17.0 // indirect
 	golang.org/x/sync v0.8.0 // indirect
 	golang.org/x/tools v0.21.1-0.20240508182429-e35e4ccd0d2d // indirect
-	gopkg.in/warnings.v0 v0.1.2 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )