--- a/go.mod
+++ b/go.mod
@@ -11,6 +11,7 @@ replace (
 
 require (
 	dario.cat/mergo v1.0.0
+	github.com/Microsoft/go-winio v0.6.1
 	github.com/ProtonMail/go-crypto v1.0.0
 	github.com/armon/go-socks5 v0.0.0-20160902184237-e75332964ef5
 	github.com/elazarl/goproxy v0.0.0-20240618083138-03be62527ccb
@@ -24,9 +25,7 @@ require (
 	github.com/kevinburke/ssh_config v1.2.0
 	github.com/pjbgf/sha1cd v0.3.0
 	github.com/sergi/go-diff v1.3.2-0.20230802210424-5b0b94c5c0d3
-	github.com/skeema/knownhosts v1.3.0
 	github.com/stretchr/testify v1.9.0
-	github.com/xanzy/ssh-agent v0.3.3
 	golang.org/x/crypto v0.26.0
 	golang.org/x/net v0.28.0
 	golang.org/x/sys v0.24.0
@@ -35,7 +34,6 @@ require (
 )
 
 require (
-	github.com/Microsoft/go-winio v0.6.1 // indirect
 	github.com/anmitsu/go-shlex v0.0.0-20200514113438-38f4b401e2be // indirect
 	github.com/cloudflare/circl v1.3.7 // indirect
 	github.com/cyphar/filepath-securejoin v0.2.4 // indirect