--- a/plumbing/format/packfile/common.go
+++ b/plumbing/format/packfile/common.go
@@ -2,9 +2,11 @@ package packfile
 
 import (
 	"io"
+	"time"
 
 	"github.com/go-git/go-git/v5/plumbing/storer"
 	"github.com/go-git/go-git/v5/utils/ioutil"
+	"github.com/go-git/go-git/v5/utils/trace"
 )
 
 var signature = []byte{'P', 'A', 'C', 'K'}
@@ -24,6 +26,11 @@ const (
 // UpdateObjectStorage updates the storer with the objects in the given
 // packfile.
 func UpdateObjectStorage(s storer.Storer, packfile io.Reader) error {
+	start := time.Now()
+	defer func() {
+		trace.Performance.Printf("performance: %.9f s: update_obj_storage", time.Since(start).Seconds())
+	}()
+
 	if pw, ok := s.(storer.PackfileWriter); ok {
 		return WritePackfileToObjectStorage(pw, packfile)
 	}