--- a/plumbing/transport/ssh/auth_method.go
+++ b/plumbing/transport/ssh/auth_method.go
@@ -3,14 +3,16 @@ package ssh
 import (
 	"errors"
 	"fmt"
+	"net"
 	"os"
 	"os/user"
 	"path/filepath"
 
 	"github.com/go-git/go-git/v5/plumbing/transport"
+	"github.com/go-git/go-git/v5/plumbing/transport/ssh/knownhosts"
+	"github.com/go-git/go-git/v5/plumbing/transport/ssh/sshagent"
+	"github.com/go-git/go-git/v5/utils/trace"
 
-	"github.com/skeema/knownhosts"
-	sshagent "github.com/xanzy/ssh-agent"
 	"golang.org/x/crypto/ssh"
 )
 
@@ -54,6 +56,7 @@ func (a *KeyboardInteractive) String() string {
 }
 
 func (a *KeyboardInteractive) ClientConfig() (*ssh.ClientConfig, error) {
+	trace.SSH.Printf("ssh: %s user=%s", KeyboardInteractiveName, a.User)
 	return a.SetHostKeyCallback(&ssh.ClientConfig{
 		User: a.User,
 		Auth: []ssh.AuthMethod{
@@ -78,6 +81,7 @@ func (a *Password) String() string {
 }
 
 func (a *Password) ClientConfig() (*ssh.ClientConfig, error) {
+	trace.SSH.Printf("ssh: %s user=%s", PasswordName, a.User)
 	return a.SetHostKeyCallback(&ssh.ClientConfig{
 		User: a.User,
 		Auth: []ssh.AuthMethod{ssh.Password(a.Password)},
@@ -101,6 +105,7 @@ func (a *PasswordCallback) String() string {
 }
 
 func (a *PasswordCallback) ClientConfig() (*ssh.ClientConfig, error) {
+	trace.SSH.Printf("ssh: %s user=%s", PasswordCallbackName, a.User)
 	return a.SetHostKeyCallback(&ssh.ClientConfig{
 		User: a.User,
 		Auth: []ssh.AuthMethod{ssh.PasswordCallback(a.Callback)},
@@ -150,6 +155,9 @@ func (a *PublicKeys) String() string {
 }
 
 func (a *PublicKeys) ClientConfig() (*ssh.ClientConfig, error) {
+	trace.SSH.Printf("ssh: %s user=%s signer=\"%s %s\"", PublicKeysName, a.User,
+		a.Signer.PublicKey().Type(),
+		ssh.FingerprintSHA256(a.Signer.PublicKey()))
 	return a.SetHostKeyCallback(&ssh.ClientConfig{
 		User: a.User,
 		Auth: []ssh.AuthMethod{ssh.PublicKeys(a.Signer)},
@@ -160,8 +168,10 @@ func username() (string, error) {
 	var username string
 	if user, err := user.Current(); err == nil {
 		username = user.Username
+		trace.SSH.Printf("ssh: Falling back to current user name %q", username)
 	} else {
 		username = os.Getenv("USER")
+		trace.SSH.Printf("ssh: Falling back to environment variable USER %q", username)
 	}
 
 	if username == "" {
@@ -211,9 +221,10 @@ func (a *PublicKeysCallback) String() string {
 }
 
 func (a *PublicKeysCallback) ClientConfig() (*ssh.ClientConfig, error) {
+	trace.SSH.Printf("ssh: %s user=%s", PublicKeysCallbackName, a.User)
 	return a.SetHostKeyCallback(&ssh.ClientConfig{
 		User: a.User,
-		Auth: []ssh.AuthMethod{ssh.PublicKeysCallback(a.Callback)},
+		Auth: []ssh.AuthMethod{tracePublicKeysCallback(a.Callback)},
 	})
 }
 
@@ -236,16 +247,17 @@ func NewKnownHostsCallback(files ...string) (ssh.HostKeyCallback, error) {
 
 func newKnownHostsDb(files ...string) (*knownhosts.HostKeyDB, error) {
 	var err error
-
 	if len(files) == 0 {
 		if files, err = getDefaultKnownHostsFiles(); err != nil {
 			return nil, err
 		}
 	}
+	trace.SSH.Printf("ssh: known_hosts sources %s", files)
 
 	if files, err = filterKnownHostsFiles(files...); err != nil {
 		return nil, err
 	}
+	trace.SSH.Printf("ssh: filtered known_hosts sources %s", files)
 
 	return knownhosts.NewDB(files...)
 }
@@ -253,6 +265,7 @@ func newKnownHostsDb(files ...string) (*knownhosts.HostKeyDB, error) {
 func getDefaultKnownHostsFiles() ([]string, error) {
 	files := filepath.SplitList(os.Getenv("SSH_KNOWN_HOSTS"))
 	if len(files) != 0 {
+		trace.SSH.Printf("ssh: loading known_hosts from SSH_KNOWN_HOSTS")
 		return files, nil
 	}
 
@@ -309,6 +322,32 @@ func (m *HostKeyCallbackHelper) SetHostKeyCallback(cfg *ssh.ClientConfig) (*ssh.
 		m.HostKeyCallback = db.HostKeyCallback()
 	}
 
-	cfg.HostKeyCallback = m.HostKeyCallback
+	cfg.HostKeyCallback = m.traceHostKeyCallback
 	return cfg, nil
+}
+
+func (m *HostKeyCallbackHelper) traceHostKeyCallback(hostname string, remote net.Addr, key ssh.PublicKey) error {
+	trace.SSH.Printf(
+		`ssh: hostkey callback hostname=%s remote=%s pubkey="%s %s"`,
+		hostname, remote, key.Type(), ssh.FingerprintSHA256(key))
+	return m.HostKeyCallback(hostname, remote, key)
+}
+
+func tracePublicKeysCallback(getSigners func() ([]ssh.Signer, error)) ssh.AuthMethod {
+	signers, err := getSigners()
+	if err != nil {
+		trace.SSH.Printf("ssh: error calling getSigners: %v", err)
+	}
+	if len(signers) == 0 {
+		trace.SSH.Printf("ssh: no signers found")
+	}
+	for _, s := range signers {
+		trace.SSH.Printf("ssh: found key: %s %s", s.PublicKey().Type(),
+			ssh.FingerprintSHA256(s.PublicKey()))
+	}
+
+	cb := func() ([]ssh.Signer, error) {
+		return signers, err
+	}
+	return ssh.PublicKeysCallback(cb)
 }