Introduces the ability to trace the SSH handshake by either using `trace.SetTarget(trace.SSH)` or via running `go-git` when the environment variable `GIT_TRACE_SSH=true` is set.

```
$ GIT_TRACE_SSH=true go run _examples/memory/main.go git@github.com:go-git/go-git
git clone git@github.com:go-git/go-git
07:14:40.065789 common.go:43: ssh: Using default auth builder (user: git)
07:14:40.065826 sshagent.go:42: ssh: net.Dial unix sock /tmp/ssh-XXXXXXHLktH8/agent.4166
07:14:40.065876 auth_method.go:224: ssh: ssh-public-key-callback user=git
07:14:40.065954 auth_method.go:345: ssh: found key: sk-ssh-ed25519@openssh.com SHA256:otsLd7qnCx9TrhUfIPCiP2cbk3Rkb5gtxyUJy2Zeheo
07:14:40.065964 auth_method.go:255: ssh: known_hosts sources [/home/user/.ssh/known_hosts /etc/ssh/ssh_known_hosts]
07:14:40.065973 auth_method.go:260: ssh: filtered known_hosts sources [/home/user/.ssh/known_hosts]
07:14:40.066288 common.go:155: ssh: host key algorithms [ssh-ed25519]
07:14:40.335435 auth_method.go:330: ssh: hostkey callback hostname=github.com:22 remote=20.26.156.215:22 pubkey="ssh-ed25519 SHA256:+DiY3wvvV6TuJJhbpZisF/zLDA0zPMSvHdkr4UvCOqU"
commit 5c762aefcd8dded79e25bc2055254b4146e2b5a9
```
This will be especially useful for users struggling with SSH auth. As per #218, #526, #699 and #489.

The same applies to Performance: 
```
$ GIT_TRACE_PERFORMANCE=true go run _examples/memory/main.go https://github.com/go-git/go-git
git clone https://github.com/go-git/go-git
07:12:47.822088 upload_pack.go:41: performance: 0.300771529 s: upload_pack
07:12:48.608808 common.go:31: performance: 0.786561842 s: update_obj_storage
07:12:48.615222 worktree.go:330: performance: 0.006302458 s: reset_worktree
07:12:48.615274 repository.go:241: performance: 1.230773190 s: git command: git clone https://github.com/go-git/go-git
commit 5c762aefcd8dded79e25bc2055254b4146e2b5a9
```

The `v6-exp` branch now points to a commit of `go-git/gcfg` in its own `v6-exp` branch, with minor [changes](https://github.com/go-git/gcfg/compare/v1...v6-exp) which includes the removal of abandoned dependencies (e.g. `gopkg.in/warnings.v0`) and the removal of `panic` calls.

cc: @aymanbagabas