@evanelias 2025-01-29
@pjbgf I know I'm 4 months late seeing this, but why does this PR copy-and-paste dependencies into new files and remove them from go.mod entirely?

Speaking as the author/maintainer of one of these dependencies (https://github.com/skeema/knownhosts), this approach raises some significant security concerns. You're now using a completely orphaned copy of the code which is otherwise disconnected from the normal Golang module flow, CVE reporting checks using `govulncheck`, etc. Not to mention, lack of normal module flow for bug fixes and enhancements.

What's the benefit for doing this in this manner?