@pjbgf 2025-01-30
> I know I'm 4 months late seeing this, but why does this PR copy-and-paste dependencies into new files and remove them from go.mod entirely?

@evanelias Hey Evan, thank you for raising your concerns, it is better late than later. :wink: On a serious note, I get where you are coming from, the changes lead to sub-optimal tracking for upstream CVEs. Please note this was merged to an experimental branch, which became the main development branch now that v5 has been [frozen](https://github.com/go-git/go-git/commit/f81dfe448c277572640d12f2968cfc2775e0aec0) for development.

The primary goal of this project is to align with the upstream Git functionality. The [changes](https://github.com/go-git/go-git/pull/1172/commits/c83efa5acd8e64ca78d455a773a2a3a54ba67503) enable a deeper level of instrumentation when tracing SSH connections, and it does so by injecting calls to `trace.SSH.Printf` in areas that are deep in the upstream code. I created an [issue](https://github.com/go-git/go-git/issues/1413) to review existing dependencies and investigate the possibility of upstreaming such changes.